+++ Example/Example_BusinessComponents/ViewController.swift
@@ -9,7 +9,8 @@
 import UIKit
 import MLBusinessComponents
 
-class ViewController: UIViewController {
+class ViewController: UIViewController, MLBusinessLoyaltyBroadcastReceiver {
+    
     @IBOutlet private weak var containerView: UIView!
     
     private weak var ringView: MLBusinessLoyaltyRingView?
@@ -18,19 +19,26 @@ class ViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        setupView()
+        setupView(self)
     }
 
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         animateRing()
     }
+    
+    func receiveInfo(_ notification: Notification) {
+        if let loyaltyBroadcastData = notification.object as? MLBusinessLoyaltyBroadcastData {
+            NSLog("Received info from broadcast: Level %i - Percentage %d - Color %@", loyaltyBroadcastData.level, loyaltyBroadcastData.percentage, loyaltyBroadcastData.primaryColor)
+        }
+     }
+    
 }
 
 // MARK: Setups
 extension ViewController {
-    private func setupView() {
-        let newRingView = setupRingView()
+    private func setupView(_ receiver: MLBusinessLoyaltyBroadcastReceiver) {
+        let newRingView = setupRingView(receiver)
         self.ringView = newRingView
         let dividingLineView = setupDividingLineView(bottomOf: newRingView)
         let itemDescriptionView = setupItemDescriptionView(bottomOf: dividingLineView)
@@ -49,8 +57,9 @@ extension ViewController {
         openSheet.bottomAnchor.constraint(equalTo: containerView.bottomAnchor, constant: -64).isActive = true
     }
 
-    private func setupRingView() -> MLBusinessLoyaltyRingView {
-        let ringView = MLBusinessLoyaltyRingView(LoyaltyRingData(), fillPercentProgress: false)
+    private func setupRingView(_ receiver: MLBusinessLoyaltyBroadcastReceiver) -> MLBusinessLoyaltyRingView {
+        let ringData = LoyaltyRingData()
+        let ringView = MLBusinessLoyaltyRingView(ringData, fillPercentProgress: false)
 
         containerView.addSubview(ringView)
 
@@ -63,7 +72,13 @@ extension ViewController {
         ringView.addTapAction { deepLink in
             print(deepLink)
         }
-
+        
+        let broadcaster = MLBusinessLoyaltyBroadcaster.instance as MLBusinessLoyaltyBroadcaster
+        broadcaster.register(receiver)
+        broadcaster.updateInfo(MLBusinessLoyaltyBroadcastData(level: ringData.getRingNumber(),
+                                                              percentage: ringData.getRingPercentage(),
+                                                              primaryColor: ringData.getRingHexaColor()))
+        
         return ringView
     }