+++ Source/Components/Loyalty/Broadcaster/MLBusinessLoyaltyBroadcastReceiver.swift
@@ -1,0 +1,12 @@
+//
+//  MLBusinessLoyaltyBroadcastReceiver.swift
+//  MLBusinessComponents
+//
+//  Created by Guido Maceira on 6/26/20.
+//
+import UIKit
+
+@objc public protocol MLBusinessLoyaltyBroadcastReceiver {
+    @objc func receiveInfo(_ notification: Notification)
+}
+