+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Sin publicar
+- Loyalty Broadcast
+
 # v1.20.0
 🚀 1.20.0 🚀
 - New Sheet Component